Create our own Sync engine using Dropbox
Trying to use the dropbox core api to sync the files. It uses the
delegate approach to requests meaning I have to wrap that up to use a
more callback style. I do this with Signal<T>. The DropboxClient
encapsulates the requests. The SyncEngine does all the syncing with the
client and uses a Store to comunicate with CoreData. Then the cascade
view controller should monitor changes on the managed object context
after doing an inital fetch request to batch update the collection view.

- Abstract the data directory from CoreData so it is calcualted on the
  fly. This is because the full path of the documents directory changes
  in the simulator.

- Abstract the ImageStore from the Store so there is a layer above the
  CoreData implementation details.

- Display static images instead of animating gifs

- Move syncing into the main VC

- Rewrite the image cache to be independent of the cells. A cell
  requests an image from the cache. If the image is in the cache, it is
  immediately returned. If the image is not in the cache, it is
  requested from disk. After the image is retrieved and added to the
  cache, a notification fires to all those who listen (all the cells
  without an image) and they can re-request for their image.