Remove Hockey and makeing archive work
It's not a part of this git commit but, Cascade has a Frameworks folder
in it's .framework directory that is automatically generated I think.
Anyways, it cannot be there to post to the store. We should figure out
how to solve this long term but for now, make sure to delete that folder
before Archiving a build.