Fix CPU utilization spiking
The CPU was spiking after displaying the initial visible images in the
collection view. This was because Cascade was asking for an image to a
cell with a size of CGSizeZero. Internally, the image fetching operation
was returning nil images after resizing to CGSizeZero and this was
causing a crash on a thread that was spiking the cpu.

We now check for the size being CGSizeZero before fetching the image and
return early if so. We may want to look deeper into Cascade to see why
it's asking for a cell that isn't visible or that is CGSizeZero.

This also abstracts the image resizing and eliminates the need for a
custom NSOperation to fetch images.