Rasterize the shadow
This greatly improves the collection view scrolling performance. I also
tried setting the shadowPath but didn't know what to set the path to in
order to make it work. I set it to the bounds of shadowView but that
caused shadow to cast over other gifs and was generally weird.