Add Facebook Messenger GIF sharing
With the release of the Facebook Messenger platform, you can now add the
SDK to your app and share gifs via that SDK. It's totally lame that
their normal share extension doesn't work for gifs but I really want
this capability so it wasn't much to add. It's kinda nice in the
Messenger app because it says the gif was posted from Poppins and even
has the icon. Then there is a reply button for anyone who was Poppins
installed that will bring you right into the app.